+++ main-server/src/main/java/com/example/demo/domain/recommendation/controller/RecommendationController.java
@@ -88,25 +88,6 @@ public class RecommendationController {
         return ResponseEntity.ok(response);
     }
 
-    @Operation(
-            summary = "팔로우 기반 추천 (레거시 예정)",
-            description = "사용자가 아직 팔로우하지 않은 최신 플레이리스트를 추천합니다."
-    )
-    @ApiResponse(
-            responseCode = "200",
-            description = "추천된 플레이리스트 카드 목록",
-            content = @Content(array = @ArraySchema(schema = @Schema(implementation = RecommendedPlaylistResponse.class)))
-    )
-    @GetMapping("/follow")
-    public ResponseEntity<List<RecommendedPlaylistResponse>> recommendFromLikedPlaylists(
-            @Parameter(hidden = true)
-            @AuthenticationPrincipal CustomUserDetails user
-    ) {
-        List<RecommendedPlaylistResponse> response = recommendationService.recommendFromLikedPlaylists(user.getId());
-        return ResponseEntity.ok(response);
-    }
-
-
     @Operation(summary = "추천 장르 목록")
     @ApiResponse(responseCode = "200", description = "추천 플레이리스트 상세 목록",
             content = @Content(array = @ArraySchema(schema = @Schema(implementation = RecommendedGenreResponse.class))))
