+++ main-server/src/main/java/com/example/demo/domain/recommendation/service/RecommendationService.java
@@ -12,8 +12,6 @@ public interface RecommendationService {
 
     List<RecommendedPlaylistResponse> getRecommendations(String userId);
 
-    List<RecommendedPlaylistResponse> recommendFromLikedPlaylists(String myUserId);
-
     List<RecommendedGenreResponse> recommendGenres();
 
     List<RecommendedPlaylistResponse> getAdminRecommendation(int limit);
